refactor: split up ParamsInjector to App & Lib injectors, rename CLASSLOADER to CLASSPATH
