+++ bro-common/src/main/java/me/xx2bab/bro/common/Constants.java
@@ -7,7 +7,7 @@ public class Constants {
     public static final String ANNO_PROC_ARG_MODULE_BUILD_TYPE = "moduleBuildType";
     public static final String ANNO_PROC_ARG_MODULE_BUILD_DIR = "moduleBroBuildDir";
     public static final String ANNO_PROC_ARG_MODULE_PROCESSOR_CLASSES = "moduleProcessorClasses";
-    public static final String ANNO_PROC_ARG_MODULE_GENERATOR_CLASSLOADERS = "moduleProcessorClassLoaders";
+    public static final String ANNO_PROC_ARG_MODULE_GENERATOR_CLASSPATHS = "moduleProcessorClassPaths";
     public static final String ANNO_PROC_ARG_APP_PACKAGE_NAME = "appPackageName";
     public static final String ANNO_PROC_ARG_APP_META_DATA_INPUT_PATH = "appMetaDataInputPath";
     public static final String ANNO_PROC_ARG_APP_APT_PATH = "appAptPath";
